Add context menus to extensions and actually register them. Rename commands to slash commands everywhere.
