+++ src/index.ts
@@ -1,3 +1,3 @@
 export { Monno, MonnoClientOptions } from "./client"
-export { MonnoCommand, MonnoCommandManager } from "./commands"
+export { MonnoCommand, MonnoSlashCommandManager } from "./slashCommands"
 export { MonnoExtension, MonnoExtensionManager } from "./extensions"
